requirements.txt
asgiref==3.2.3
Django==1.11.28
django-filter==2.2.0
djangorestframework==3.9.4
djangorestframework-filters==1.0.0.dev0
Markdown==3.1.1
pytz==2019.3
six==1.14.0
sqlparse==0.3.0
